Rename files, fix demo file
