+++ demo/d2l-quick-eval-widget/d2l-quick-eval-widget.html
@@ -23,6 +23,8 @@
 				token=""
 				href=""
 				quick-eval-href=""
+				toggle-href=""
+				toggle-state="activities"
 				count="3">
 				</d2l-quick-eval-widget>
 			</d2l-demo-snippet>
